--- kfp/transform_workflows/code/Makefile
@@ -1,29 +1,0 @@
-REPOROOT=../../../
-# Use "make help" to see them.
-include ../../../.make.defaults
-
-clean:: 
-	@# Help: Recursively $@ in all subdirs 
-	$(MAKE) RULE=$@ .recurse
-
-setup::
-
-build::
-	$(MAKE) RULE=$@ .recurse
-
-venv:: 
-	@# Help: Recursively $@ in all subdirs
-	@$(MAKE) RULE=$@ .recurse
-
-test:: 
-	@# Help: Recursively make $@ in all subdirs 
-	@$(MAKE) RULE=$@ .recurse
-
-upload::
-	@# Help: Recursively make $@ in all subdirs
-	@$(MAKE) RULE=$@ .recurse
-
-reconcile-requirements::
-	@# Help: Recursively make $@ in all subdirs
-	@$(MAKE) RULE=$@ .recurse
-
