+++ transforms/.make.transforms_workflows
@@ -1,0 +1,66 @@
+include ${REPOROOT}/.make.versions
+include ${REPOROOT}/kfp/requirements.env
+
+# Include the common rules.
+# Use "make help" to see them.
+include ${REPOROOT}/.make.defaults
+
+USE_DEV_IMAGES ?= 1
+
+define set_env_var
+	$(eval export $(1)=$(2))
+endef
+
+.PHONY: .transforms_workflows.reconcile-requirement
+.transforms_workflows.reconcile-requirement:
+	@while IFS= read -r line; do \
+		[ -z "$$line" ] && continue; \
+		[[ $$line == *#* ]] && continue; \
+		export DOCKER_IMAGE_NAME=$$(echo $$line |cut -d "=" -f 1 |sed "s/_VERSION//" |tr '[:upper:]' '[:lower:]'); \
+		export DOCKER_IMAGE_VERSION=$$(echo $$line |cut -d "=" -f 2); \
+		sed -i.back "s/data-prep-kit\/$$DOCKER_IMAGE_NAME:.*/data-prep-kit\/$$DOCKER_IMAGE_NAME:$$DOCKER_IMAGE_VERSION\"/" $$PIPELINE_FILE ;\
+	done < ${REPOROOT}/.make.versions
+	@sed -i.back "s/kfp-data-processing:.*/kfp-data-processing:${KFP_DOCKER_VERSION}\"/" ${PIPELINE_FILE}
+
+.PHONY: .transforms_workflows.compile-pipeline
+.transforms_workflows.compile-pipeline:
+	. ${WORKFLOW_VENV_ACTIVATE} && ${PYTHON} ${WF_NAME}.py
+
+%.yaml: %.py ${REPOROOT}/.make.versions ${REPOROOT}/kfp/requirements.env ${DPK_RAY_LIB_DIR} ${REPOROOT}/kfp/kfp_support_lib/
+	$(MAKE) .transforms_workflows.reconcile-requirement PIPELINE_FILE=$<
+	$(MAKE) .transforms_workflows.compile-pipeline WF_NAME=$(shell (basename $< .py))
+
+.PHONY: .transforms_workflows.test-pipeline
+.transforms_workflows.test-pipeline:
+	@# Help: upload and run the workflow. Set export USE_DEV_IMAGES=0 to use release docker image versions.
+	$(call set_env_var, CLUSTER_EXISTS, $(shell kind get clusters | grep ${KIND_CLUSTER_NAME}))
+	@if [ -z ${CLUSTER_EXISTS} ]; then \
+		cd ${REPOROOT} && make setup;  \
+	fi
+ifeq ($(USE_DEV_IMAGES), 1)
+	cd ${TRANSFORM_SRC} && $(MAKE) image && $(MAKE) load-image
+	cd ${REPOROOT}/kfp/kfp_ray_components && $(MAKE) image && $(MAKE) load-image
+endif
+	. ${WORKFLOW_VENV_ACTIVATE}  && ${PYTHON} -m kfp_support.workflow_support.utils.pipelines_tests_utils -c "sanity-test" -p ${CURDIR}/${PIPELINE_FILE}
+
+${WORKFLOW_VENV_ACTIVATE}: ${REPOROOT}/.make.versions ${REPOROOT}/kfp/requirements.env ${REPOROOT}/kfp/kfp_ray_components/requirements.txt ${DPK_RAY_LIB_DIR} ${REPOROOT}/kfp/kfp_support_lib/
+	@# Help: Create the virtual environment common to all workflows
+	rm -rf ${REPOROOT}/transforms/venv
+	$(PYTHON) -m venv ${REPOROOT}/transforms/venv
+	. ${WORKFLOW_VENV_ACTIVATE};     \
+	pip install kfp==${KFP} --extra-index-url https://pypi.org/simple; \
+	pip install -e $($DPK_RAY_LIB_DIR); \
+	pip install -e $(REPOROOT)/kfp/kfp_support_lib/;
+	
+.PHONY: .transforms_workflows.upload-pipeline
+.transforms_workflows.upload-pipeline:
+	$(call set_env_var, CLUSTER_EXISTS, $(shell kind get clusters | grep ${KIND_CLUSTER_NAME}))
+	@if [ -z ${CLUSTER_EXISTS} ]; then \
+		cd ${REPOROOT} && make setup;  \
+	fi
+	. ${WORKFLOW_VENV_ACTIVATE}  && ${PYTHON} -m kfp_support.workflow_support.utils.pipelines_tests_utils -c "upload" -p ${CURDIR}/${PIPELINE_FILE}
+
+.PHONY: .transforms_workflows.clean
+.transforms_workflows.clean:
+	rm -rf *_wf.yaml || true
+
