+++ transforms/code/malware/Makefile
@@ -40,3 +40,19 @@ test-src::
 load-image::
 	@# Help: Recursively make $@ in all subdirs 
 	$(MAKE) RULE=$@ .recurse
+
+workflow-venv:
+	@# Help: Recursively make $@ in subdirs
+	$(MAKE) RULE=$@ .recurse	
+
+workflow-build:
+	@# Help: Recursively make $@ in subdirs
+	$(MAKE) RULE=$@ .recurse
+
+workflow-test:
+	@# Help: Recursively make $@ in subdirs
+	$(MAKE) RULE=$@ .recurse
+
+workflow-upload:
+	@# Help: Recursively make $@ in subdirs
+	$(MAKE) RULE=$@ .recurse
