+++ transforms/code/malware/ray/kfp-workflow/malware_wf.py
@@ -21,39 +21,43 @@ from kfp_support.workflow_support.utils import (
 
 
 # the name of the job script
-EXEC_SCRIPT_NAME: str = "tokenization_transform.py"
+EXEC_SCRIPT_NAME: str = "malware_transform.py"
 
-task_image = "quay.io/dataprep1/data-prep-kit/tokenization:0.2.0"
+task_image = "quay.io/dataprep1/data-prep-kit/malware:0.4.0"
 
 # components
 base_kfp_image = "quay.io/dataprep1/data-prep-kit/kfp-data-processing:0.1.0"
-# compute execution parameters. Use default one for now.
+
+# path to kfp component specifications files
+component_spec_path = "../../../../../kfp/kfp_ray_components/"
+
+# compute execution parameters. Here different tranforms might need different implementations. As
+# a result, insted of creating a component we are creating it in place here.
 compute_exec_params_op = comp.func_to_container_op(
     func=ComponentUtils.default_compute_execution_params, base_image=base_kfp_image
 )
 # create Ray cluster
-create_ray_op = comp.load_component_from_file("../../../kfp_ray_components/createRayComponent.yaml")
+create_ray_op = comp.load_component_from_file(component_spec_path + "createRayComponent.yaml")
 # execute job
-execute_ray_jobs_op = comp.load_component_from_file("../../../kfp_ray_components/executeRayJobComponent.yaml")
+execute_ray_jobs_op = comp.load_component_from_file(component_spec_path + "executeRayJobComponent.yaml")
 # clean up Ray
-cleanup_ray_op = comp.load_component_from_file("../../../kfp_ray_components/cleanupRayComponent.yaml")
+cleanup_ray_op = comp.load_component_from_file(component_spec_path + "cleanupRayComponent.yaml")
 # Task name is part of the pipeline name, the ray cluster name and the job name in DMF.
-TASK_NAME: str = "tokenization"
+TASK_NAME: str = "malware"
 
 
 @dsl.pipeline(
     name=TASK_NAME + "-ray-pipeline",
-    description="Pipeline for tokenization",
+    description="Pipeline for malware",
 )
-def tokenization(
-    # Ray cluster
-    ray_name: str = "tkn-kfp-ray",  # name of Ray cluster
+def malware(
+    ray_name: str = "malware-kfp-ray",  # name of Ray cluster
     ray_head_options: str = '{"cpu": 1, "memory": 4, "image_pull_secret": "", "image": "' + task_image + '" }',
     ray_worker_options: str = '{"replicas": 2, "max_replicas": 2, "min_replicas": 2, "cpu": 2, "memory": 4, '
     '"image_pull_secret": "", "image": "' + task_image + '"}',
     server_url: str = "http://kuberay-apiserver-service.kuberay.svc.cluster.local:8888",
     # data access
-    data_s3_config: str = "{'input_folder': 'test/tokenization/ds01/input/', 'output_folder': 'test/tokenization/ds01/output/'}",
+    data_s3_config: str = "{'input_folder': 'test/malware/input', 'output_folder': 'test/malware/output'}",
     data_s3_access_secret: str = "s3-secret",
     data_max_files: int = -1,
     data_num_samples: int = -1,
@@ -61,18 +65,14 @@ def tokenization(
     runtime_actor_options: str = "{'num_cpus': 0.8}",
     runtime_pipeline_id: str = "pipeline_id",
     runtime_code_location: str = "{'github': 'github', 'commit_hash': '12345', 'path': 'path'}",
-    # tokenizer parameters
-    tkn_tokenizer: str = "hf-internal-testing/llama-tokenizer",
-    tkn_doc_id_column: str = "document_id",
-    tkn_doc_content_column: str = "contents",
-    tkn_text_lang: str = "en",
-    tkn_tokenizer_args: str = "cache_dir=/tmp/hf",
-    tkn_chunk_size: int = 0,
+    # mallware
+    malware_input_column: str = "contents",
+    malware_output_column: str = "virus_detection",
     # additional parameters
     additional_params: str = '{"wait_interval": 2, "wait_cluster_ready_tmout": 400, "wait_cluster_up_tmout": 300, "wait_job_ready_tmout": 400, "wait_print_tmout": 30, "http_retries": 5}',
 ):
     """
-    Pipeline to execute tokenization transform
+    Pipeline to execute malware transform
     :param ray_name: name of the Ray cluster
     :param ray_head_options: head node options, containing the following:
         cpu - number of cpus
@@ -94,20 +94,16 @@ def tokenization(
         wait_cluster_up_tmout - time to wait for cluster up, sec
         wait_job_ready_tmout - time to wait for job ready, sec
         wait_print_tmout - time between prints, sec
-        http_retries - http retries for API server calls
-    :param data_s3_access_secret - s3 access secret
+        http_retries - httpt retries for API server calls
     :param data_s3_config - s3 configuration
+    :param data_s3_access_secret - s3 access secret
     :param data_max_files - max files to process
     :param data_num_samples - num samples to process
     :param runtime_actor_options - actor options
     :param runtime_pipeline_id - pipeline id
     :param runtime_code_location - code location
-    :param tkn_tokenizer - Tokenizer used for tokenization
-    :param tkn_tokenizer_args - Arguments for tokenizer.
-    :param tkn_doc_id_column - Column contains document id which values should be unique across dataset
-    :param tkn_doc_content_column - Column contains document content
-    :param tkn_text_lang - Specify language used in the text content for better text splitting if needed
-    :param tkn_chunk_size - Specify >0 value to tokenize each row/text in chunks of characters (rounded in words)
+    :param malware_input_column - input column name
+    :param malware_output_column - output column name
     :return: None
     """
     # create clean_up task
@@ -121,8 +117,6 @@ def tokenization(
             actor_options=runtime_actor_options,
         )
         ComponentUtils.add_settings_to_component(compute_exec_params, ONE_HOUR_SEC * 2)
-        ComponentUtils.set_s3_env_vars_to_component(compute_exec_params, data_s3_access_secret)
-
         # start Ray cluster
         ray_cluster = create_ray_op(
             ray_name=ray_name,
@@ -139,6 +133,7 @@ def tokenization(
             ray_name=ray_name,
             run_id=dsl.RUN_ID_PLACEHOLDER,
             additional_params=additional_params,
+            # note that the parameters below are specific for malware transform
             exec_params={
                 "data_s3_config": data_s3_config,
                 "data_max_files": data_max_files,
@@ -148,12 +143,8 @@ def tokenization(
                 "runtime_pipeline_id": runtime_pipeline_id,
                 "runtime_job_id": dsl.RUN_ID_PLACEHOLDER,
                 "runtime_code_location": runtime_code_location,
-                "tkn_tokenizer": tkn_tokenizer,
-                "tkn_tokenizer_args": tkn_tokenizer_args,
-                "tkn_doc_id_column": tkn_doc_id_column,
-                "tkn_doc_content_column": tkn_doc_content_column,
-                "tkn_text_lang": tkn_text_lang,
-                "tkn_chunk_size": tkn_chunk_size,
+                "malware_input_column": malware_input_column,
+                "malware_output_column": malware_output_column,
             },
             exec_script_name=EXEC_SCRIPT_NAME,
             server_url=server_url,
@@ -168,4 +159,4 @@ def tokenization(
 
 if __name__ == "__main__":
     # Compiling the pipeline
-    compiler.Compiler().compile(tokenization, __file__.replace(".py", ".yaml"))
+    compiler.Compiler().compile(malware, __file__.replace(".py", ".yaml"))
