+++ transforms/code/proglang_select/ray/kfp-workflow/proglang_select_wf.py
@@ -10,8 +10,6 @@
 # limitations under the License.
 ################################################################################
 
-# NOTE: This file is auto generated by Pipeline Generator.
-
 import kfp.compiler as compiler
 import kfp.components as comp
 import kfp.dsl as dsl
@@ -20,68 +18,62 @@ from kfp_support.workflow_support.utils import (
     ONE_WEEK_SEC,
     ComponentUtils,
 )
-from kubernetes import client as k8s_client
 
 
 # the name of the job script
-EXEC_SCRIPT_NAME: str = "code_quality_transform.py"
-PREFIX: str = ""
+EXEC_SCRIPT_NAME: str = "proglang_select_transform.py"
 
-task_image = "quay.io/dataprep1/data-prep-kit/code_quality:0.3.0"
+task_image = "quay.io/dataprep1/data-prep-kit/proglang_select:0.3.0"
 
 # components
 base_kfp_image = "quay.io/dataprep1/data-prep-kit/kfp-data-processing:0.1.0"
+
+# path to kfp component specifications files
+component_spec_path = "../../../../../kfp/kfp_ray_components/"
+
 # compute execution parameters. Here different tranforms might need different implementations. As
 # a result, insted of creating a component we are creating it in place here.
 compute_exec_params_op = comp.func_to_container_op(
     func=ComponentUtils.default_compute_execution_params, base_image=base_kfp_image
 )
 # create Ray cluster
-create_ray_op = comp.load_component_from_file("../../../kfp_ray_components/createRayComponent.yaml")
+create_ray_op = comp.load_component_from_file(component_spec_path + "createRayComponent.yaml")
 # execute job
-execute_ray_jobs_op = comp.load_component_from_file("../../../kfp_ray_components/executeRayJobComponent.yaml")
+execute_ray_jobs_op = comp.load_component_from_file(component_spec_path + "executeRayJobComponent.yaml")
 # clean up Ray
-cleanup_ray_op = comp.load_component_from_file("../../../kfp_ray_components/cleanupRayComponent.yaml")
+cleanup_ray_op = comp.load_component_from_file(component_spec_path + "cleanupRayComponent.yaml")
 # Task name is part of the pipeline name, the ray cluster name and the job name in DMF.
-TASK_NAME: str = "code_quality"
+TASK_NAME: str = "proglang_select"
 
 
-# Pipeline to invoke execution on remote resource
 @dsl.pipeline(
     name=TASK_NAME + "-ray-pipeline",
-    description="Pipeline for code quality task",
+    description="Pipeline for select language",
 )
-def code_quality(
-    # Ray cluster
-    ray_name: str = "code_quality-kfp-ray",  # name of Ray cluster
-    ray_head_options: str = '{"cpu": 1, "memory": 4, "image_pull_secret": "",\
-            "image": "'
-    + task_image
-    + '" }',
-    ray_worker_options: str = '{"replicas": 2, "max_replicas": 2, "min_replicas": 2, "cpu": 2, "memory": 4, "image_pull_secret": "",\
-            "image": "'
-    + task_image
-    + '" }',
+def lang_select(
+    ray_name: str = "proglang-match-kfp-ray",  # name of Ray cluster
+    ray_head_options: str = '{"cpu": 1, "memory": 4, "image_pull_secret": "", "image": "' + task_image + '" }',
+    ray_worker_options: str = '{"replicas": 2, "max_replicas": 2, "min_replicas": 2, "cpu": 2, "memory": 4, '
+    '"image_pull_secret": "", "image": "' + task_image + '"}',
     server_url: str = "http://kuberay-apiserver-service.kuberay.svc.cluster.local:8888",
     # data access
-    data_s3_config: str = "{'input_folder': 'test/code_quality/input/', 'output_folder': 'test/code_quality/output/'}",
+    data_s3_config: str = "{'input_folder': 'test/proglang_select/input/', 'output_folder': 'test/proglang_select/output/'}",
     data_s3_access_secret: str = "s3-secret",
     data_max_files: int = -1,
     data_num_samples: int = -1,
     # orchestrator
     runtime_actor_options: str = "{'num_cpus': 0.8}",
-    runtime_pipeline_id: str = "runtime_pipeline_id",
+    runtime_pipeline_id: str = "pipeline_id",
     runtime_code_location: str = "{'github': 'github', 'commit_hash': '12345', 'path': 'path'}",
-    # code quality parameters
-    cq_contents_column_name: str = "contents",
-    cq_language_column_name: str = "language",
-    cq_tokenizer: str = "codeparrot/codeparrot",
-    cq_hf_token: str = "None",
+    # Proglang match parameters
+    proglang_select_allowed_langs_file: str = "test/proglang_select/languages/allowed-code-languages.txt",
+    proglang_select_language_column: str = "language",
+    proglang_select_s3_access_secret: str = "s3-secret",
     # additional parameters
     additional_params: str = '{"wait_interval": 2, "wait_cluster_ready_tmout": 400, "wait_cluster_up_tmout": 300, "wait_job_ready_tmout": 400, "wait_print_tmout": 30, "http_retries": 5}',
-):
+) -> None:
     """
-    Pipeline to execute Code Quality transform
+    Pipeline to execute NOOP transform
     :param ray_name: name of the Ray cluster
     :param ray_head_options: head node options, containing the following:
         cpu - number of cpus
@@ -103,17 +95,18 @@ def code_quality(
         wait_cluster_up_tmout - time to wait for cluster up, sec
         wait_job_ready_tmout - time to wait for job ready, sec
         wait_print_tmout - time between prints, sec
-        http_retries - http retries for API server calls
+        http_retries - httpt retries for API server calls
     :param data_s3_access_secret - s3 access secret
     :param data_s3_config - s3 configuration
     :param data_max_files - max files to process
     :param data_num_samples - num samples to process
     :param runtime_actor_options - actor options
     :param runtime_pipeline_id - pipeline id
-    :param cq_contents_column_name - Name of the column holds the data to process
-    :param cq_language_column_name - Name of the column holds the programming language details
-    :param cq_tokenizer - Name or path to the tokenizer
-    :param cq_hf_token - Huggingface auth token to download and use the tokenizer
+    :param runtime_code_location - code location
+    :param proglang_select_allowed_langs_file - file to store allowed languages
+    :param proglang_select_language_column - name of select language annotation column
+    :param proglang_select_s3_access_secret - block list access secret
+                    (here we are assuming that select language info is in S3, but potentially in the different bucket)
     :return: None
     """
     # create clean_up task
@@ -138,12 +131,12 @@ def code_quality(
         )
         ComponentUtils.add_settings_to_component(ray_cluster, ONE_HOUR_SEC * 2)
         ray_cluster.after(compute_exec_params)
-
         # Execute job
         execute_job = execute_ray_jobs_op(
             ray_name=ray_name,
             run_id=dsl.RUN_ID_PLACEHOLDER,
             additional_params=additional_params,
+            # note that the parameters below are specific for NOOP transform
             exec_params={
                 "data_s3_config": data_s3_config,
                 "data_max_files": data_max_files,
@@ -152,17 +145,16 @@ def code_quality(
                 "runtime_worker_options": runtime_actor_options,
                 "runtime_pipeline_id": runtime_pipeline_id,
                 "runtime_job_id": dsl.RUN_ID_PLACEHOLDER,
-                "cq_contents_column_name": cq_contents_column_name,
-                "cq_language_column_name": cq_language_column_name,
-                "cq_tokenizer": cq_tokenizer,
-                "cq_hf_token": cq_hf_token,
+                "runtime_code_location": runtime_code_location,
+                "proglang_select_allowed_langs_file": proglang_select_allowed_langs_file,
+                "proglang_select_language_column": proglang_select_language_column,
             },
             exec_script_name=EXEC_SCRIPT_NAME,
             server_url=server_url,
         )
         ComponentUtils.add_settings_to_component(execute_job, ONE_WEEK_SEC)
         ComponentUtils.set_s3_env_vars_to_component(execute_job, data_s3_access_secret)
-
+        ComponentUtils.set_s3_env_vars_to_component(execute_job, proglang_select_s3_access_secret)
         execute_job.after(ray_cluster)
 
     # Configure the pipeline level to one week (in seconds)
@@ -171,4 +163,4 @@ def code_quality(
 
 if __name__ == "__main__":
     # Compiling the pipeline
-    compiler.Compiler().compile(code_quality, __file__.replace(".py", ".yaml"))
+    compiler.Compiler().compile(lang_select, __file__.replace(".py", ".yaml"))
