+++ transforms/universal/ededup/ray/kfp-workflow/ededup_wf.py
@@ -18,44 +18,46 @@ from kfp_support.workflow_support.utils import (
     ONE_WEEK_SEC,
     ComponentUtils,
 )
+from src.ededup_compute_execution_params import ededup_compute_execution_params
 
 
 # the name of the job script
-EXEC_SCRIPT_NAME: str = "proglang_select_transform.py"
+EXEC_SCRIPT_NAME: str = "ededup_transform.py"
 
-task_image = "quay.io/dataprep1/data-prep-kit/proglang_select:0.3.0"
+task_image = "quay.io/dataprep1/data-prep-kit/ededup:0.3.0"
 
 # components
 base_kfp_image = "quay.io/dataprep1/data-prep-kit/kfp-data-processing:0.1.0"
 
-# compute execution parameters. Here different tranforms might need different implementations. As
-# a result, insted of creating a component we are creating it in place here.
-compute_exec_params_op = comp.func_to_container_op(
-    func=ComponentUtils.default_compute_execution_params, base_image=base_kfp_image
-)
+# path to kfp component specifications files
+component_spec_path = "../../../../../kfp/kfp_ray_components/"
+
+# compute execution parameters
+compute_exec_params_op = comp.func_to_container_op(func=ededup_compute_execution_params, base_image=base_kfp_image)
 # create Ray cluster
-create_ray_op = comp.load_component_from_file("../../../kfp_ray_components/createRayComponent.yaml")
+create_ray_op = comp.load_component_from_file(component_spec_path + "createRayComponent.yaml")
 # execute job
-execute_ray_jobs_op = comp.load_component_from_file("../../../kfp_ray_components/executeRayJobComponent_multi_s3.yaml")
+execute_ray_jobs_op = comp.load_component_from_file(component_spec_path + "executeRayJobComponent.yaml")
 # clean up Ray
-cleanup_ray_op = comp.load_component_from_file("../../../kfp_ray_components/cleanupRayComponent.yaml")
+cleanup_ray_op = comp.load_component_from_file(component_spec_path + "cleanupRayComponent.yaml")
+
 # Task name is part of the pipeline name, the ray cluster name and the job name in DMF.
-TASK_NAME: str = "proglang_select"
-PREFIX: str = "proglang_select"
+TASK_NAME: str = "ededup"
 
 
 @dsl.pipeline(
     name=TASK_NAME + "-ray-pipeline",
-    description="Pipeline for select language",
+    description="Pipeline for ededup",
 )
-def lang_select(
-    ray_name: str = "proglang-match-kfp-ray",  # name of Ray cluster
+def ededup(
+    # Ray cluster
+    ray_name: str = "ededup-kfp-ray",  # name of Ray cluster
     ray_head_options: str = '{"cpu": 1, "memory": 4, "image_pull_secret": "", "image": "' + task_image + '" }',
     ray_worker_options: str = '{"replicas": 2, "max_replicas": 2, "min_replicas": 2, "cpu": 2, "memory": 4, '
     '"image_pull_secret": "", "image": "' + task_image + '"}',
     server_url: str = "http://kuberay-apiserver-service.kuberay.svc.cluster.local:8888",
-    # data access
-    data_s3_config: str = "{'input_folder': 'test/proglang_select/input/', 'output_folder': 'test/proglang_select/output/'}",
+    # data access. checkpointing is not supported by dedup
+    data_s3_config: str = "{'input_folder': 'test/ededup/input/', 'output_folder': 'test/ededup/output'}",
     data_s3_access_secret: str = "s3-secret",
     data_max_files: int = -1,
     data_num_samples: int = -1,
@@ -63,15 +65,16 @@ def lang_select(
     runtime_actor_options: str = "{'num_cpus': 0.8}",
     runtime_pipeline_id: str = "pipeline_id",
     runtime_code_location: str = "{'github': 'github', 'commit_hash': '12345', 'path': 'path'}",
-    # Proglang match parameters
-    proglang_select_allowed_langs_file: str = "test/proglang_select/languages/allowed-code-languages.txt",
-    proglang_select_language_column: str = "language",
-    proglang_select_s3_access_secret: str = "s3-secret",
+    # ededup
+    ededup_hash_cpu: float = 0.5,
+    ededup_doc_column: str = "contents",
+    # data sampling
+    ededup_n_samples: int = 10,
     # additional parameters
     additional_params: str = '{"wait_interval": 2, "wait_cluster_ready_tmout": 400, "wait_cluster_up_tmout": 300, "wait_job_ready_tmout": 400, "wait_print_tmout": 30, "http_retries": 5}',
-) -> None:
+):
     """
-    Pipeline to execute NOOP transform
+    Pipeline to execute EDEDUP transform
     :param ray_name: name of the Ray cluster
     :param ray_head_options: head node options, containing the following:
         cpu - number of cpus
@@ -93,7 +96,7 @@ def lang_select(
         wait_cluster_up_tmout - time to wait for cluster up, sec
         wait_job_ready_tmout - time to wait for job ready, sec
         wait_print_tmout - time between prints, sec
-        http_retries - httpt retries for API server calls
+        http_retries - http retries for API server calls
     :param data_s3_access_secret - s3 access secret
     :param data_s3_config - s3 configuration
     :param data_max_files - max files to process
@@ -101,10 +104,9 @@ def lang_select(
     :param runtime_actor_options - actor options
     :param runtime_pipeline_id - pipeline id
     :param runtime_code_location - code location
-    :param proglang_select_allowed_langs_file - file to store allowed languages
-    :param proglang_select_language_column - name of select language annotation column
-    :param proglang_select_s3_access_secret - block list access secret
-                    (here we are assuming that select language info is in S3, but potentially in the different bucket)
+    :param ededup_hash_cpu - number of CPUs per hash
+    :param ededup_doc_column - key for accessing data
+    :param ededup_n_samples - number of samples for parameters computation
     :return: None
     """
     # create clean_up task
@@ -116,8 +118,12 @@ def lang_select(
         compute_exec_params = compute_exec_params_op(
             worker_options=ray_worker_options,
             actor_options=runtime_actor_options,
+            params={"s3_config": data_s3_config, "hash_cpu": ededup_hash_cpu},
+            n_samples=ededup_n_samples,
         )
         ComponentUtils.add_settings_to_component(compute_exec_params, ONE_HOUR_SEC * 2)
+        ComponentUtils.set_s3_env_vars_to_component(compute_exec_params, data_s3_access_secret)
+
         # start Ray cluster
         ray_cluster = create_ray_op(
             ray_name=ray_name,
@@ -134,26 +140,24 @@ def lang_select(
             ray_name=ray_name,
             run_id=dsl.RUN_ID_PLACEHOLDER,
             additional_params=additional_params,
-            # note that the parameters below are specific for NOOP transform
             exec_params={
                 "data_s3_config": data_s3_config,
                 "data_max_files": data_max_files,
                 "data_num_samples": data_num_samples,
-                "runtime_num_workers": compute_exec_params.output,
+                "runtime_num_workers": compute_exec_params.outputs["workers"],
                 "runtime_worker_options": runtime_actor_options,
                 "runtime_pipeline_id": runtime_pipeline_id,
                 "runtime_job_id": dsl.RUN_ID_PLACEHOLDER,
                 "runtime_code_location": runtime_code_location,
-                "proglang_select_allowed_langs_file": proglang_select_allowed_langs_file,
-                "proglang_select_language_column": proglang_select_language_column,
+                "ededup_doc_column": ededup_doc_column,
+                "ededup_hash_cpu": ededup_hash_cpu,
+                "ededup_num_hashes": compute_exec_params.outputs["hashes"],
             },
             exec_script_name=EXEC_SCRIPT_NAME,
             server_url=server_url,
-            prefix=PREFIX,
         )
         ComponentUtils.add_settings_to_component(execute_job, ONE_WEEK_SEC)
         ComponentUtils.set_s3_env_vars_to_component(execute_job, data_s3_access_secret)
-        ComponentUtils.set_s3_env_vars_to_component(execute_job, proglang_select_s3_access_secret, prefix=PREFIX)
         execute_job.after(ray_cluster)
 
     # Configure the pipeline level to one week (in seconds)
@@ -162,4 +166,4 @@ def lang_select(
 
 if __name__ == "__main__":
     # Compiling the pipeline
-    compiler.Compiler().compile(lang_select, __file__.replace(".py", ".yaml"))
+    compiler.Compiler().compile(ededup, __file__.replace(".py", ".yaml"))
