+++ transforms/universal/fdedup/ray/kfp-workflow/fdedup_wf.py
@@ -10,8 +10,6 @@
 # limitations under the License.
 ################################################################################
 
-# NOTE: This file is auto generated by Pipeline Generator.
-
 import kfp.compiler as compiler
 import kfp.components as comp
 import kfp.dsl as dsl
@@ -20,45 +18,46 @@ from kfp_support.workflow_support.utils import (
     ONE_WEEK_SEC,
     ComponentUtils,
 )
+from src.fdedup_compute_execution_params import fdedup_compute_execution_params
 
 
 # the name of the job script
-EXEC_SCRIPT_NAME: str = "filter_transform.py"
-PREFIX: str = ""
+EXEC_SCRIPT_NAME: str = "fdedup_transform.py"
 
-task_image = "quay.io/dataprep1/data-prep-kit/filter:0.3.0"
+task_image = "quay.io/dataprep1/data-prep-kit/fdedup:0.3.0"
 
 # components
 base_kfp_image = "quay.io/dataprep1/data-prep-kit/kfp-data-processing:0.1.0"
-# compute execution parameters. Here different tranforms might need different implementations. As
-# a result, insted of creating a component we are creating it in place here.
-compute_exec_params_op = comp.func_to_container_op(
-    func=ComponentUtils.default_compute_execution_params, base_image=base_kfp_image
-)
+
+# path to kfp component specifications files
+component_spec_path = "../../../../../kfp/kfp_ray_components/"
+
+# compute execution parameters
+compute_exec_params_op = comp.func_to_container_op(func=fdedup_compute_execution_params, base_image=base_kfp_image)
 # create Ray cluster
-create_ray_op = comp.load_component_from_file("../../../kfp_ray_components/createRayComponent.yaml")
+create_ray_op = comp.load_component_from_file(component_spec_path + "createRayComponent.yaml")
 # execute job
-execute_ray_jobs_op = comp.load_component_from_file("../../../kfp_ray_components/executeRayJobComponent.yaml")
+execute_ray_jobs_op = comp.load_component_from_file(component_spec_path + "executeRayJobComponent.yaml")
 # clean up Ray
-cleanup_ray_op = comp.load_component_from_file("../../../kfp_ray_components/cleanupRayComponent.yaml")
+cleanup_ray_op = comp.load_component_from_file(component_spec_path + "cleanupRayComponent.yaml")
+
 # Task name is part of the pipeline name, the ray cluster name and the job name in DMF.
-TASK_NAME: str = "filter"
+TASK_NAME: str = "fdedup"
 
 
-# Pipeline to invoke execution on remote resource
 @dsl.pipeline(
     name=TASK_NAME + "-ray-pipeline",
-    description="Pipeline for filtering task",
+    description="Pipeline for fdedup",
 )
-def filtering(
+def fdedup(
     # Ray cluster
-    ray_name: str = "filter-kfp-ray",  # name of Ray cluster
+    ray_name: str = "fdedup-kfp-ray",  # name of Ray cluster
     ray_head_options: str = '{"cpu": 1, "memory": 4, "image_pull_secret": "", "image": "' + task_image + '" }',
     ray_worker_options: str = '{"replicas": 2, "max_replicas": 2, "min_replicas": 2, "cpu": 2, "memory": 4, '
     '"image_pull_secret": "", "image": "' + task_image + '"}',
     server_url: str = "http://kuberay-apiserver-service.kuberay.svc.cluster.local:8888",
-    # data access
-    data_s3_config: str = "{'input_folder': 'test/filter/input/', 'output_folder': 'test/filter/output/'}",
+    # data access. checkpointing is not supported by dedup
+    data_s3_config: str = "{'input_folder': 'test/fdedup/input/', 'output_folder': 'test/fdedup/output/'}",
     data_s3_access_secret: str = "s3-secret",
     data_max_files: int = -1,
     data_num_samples: int = -1,
@@ -66,15 +65,32 @@ def filtering(
     runtime_actor_options: str = "{'num_cpus': 0.8}",
     runtime_pipeline_id: str = "pipeline_id",
     runtime_code_location: str = "{'github': 'github', 'commit_hash': '12345', 'path': 'path'}",
-    # filtering parameters
-    filter_criteria_list: str = "['docq_total_words > 100 AND docq_total_words < 200', 'ibmkenlm_docq_perplex_score < 230']",
-    filter_logical_operator: str = "AND",
-    filter_columns_to_drop: str = "['extra', 'cluster']",
+    # columns used
+    fdedup_doc_column: str = "contents",
+    fdedup_id_column: str = "int_id_column",
+    fdedup_cluster_column: str = "cluster",
+    # infrastructure
+    fdedup_bucket_cpu: float = 0.5,
+    fdedup_doc_cpu: float = 0.5,
+    fdedup_mhash_cpu: float = 0.5,
+    # fuzzy parameters
+    fdedup_num_permutations: int = 64,
+    fdedup_threshold: float = 0.8,
+    fdedup_shingles_size: int = 5,
+    fdedup_delimiters: str = " ",
+    # Random delay between reads
+    fdedup_random_delay_limit: int = 5,
+    # snapshotting
+    fdedup_snapshot_delay: int = 1,
+    fdedup_use_doc_snapshot: bool = False,
+    fdedup_use_bucket_snapshot: bool = False,
+    # data sampling
+    fdedup_n_samples: int = 10,
     # additional parameters
     additional_params: str = '{"wait_interval": 2, "wait_cluster_ready_tmout": 400, "wait_cluster_up_tmout": 300, "wait_job_ready_tmout": 400, "wait_print_tmout": 30, "http_retries": 5}',
 ):
     """
-    Pipeline to execute Filtering transform
+    Pipeline to execute FDEDUP transform
     :param ray_name: name of the Ray cluster
     :param ray_head_options: head node options, containing the following:
         cpu - number of cpus
@@ -104,9 +120,22 @@ def filtering(
     :param runtime_actor_options - actor options
     :param runtime_pipeline_id - pipeline id
     :param runtime_code_location - code location
-    :param filter_criteria_list - list of filter criteria (in SQL WHERE clause format)
-    :param filter_logical_operator - logical operator (AND or OR) that joins filter criteria
-    :param filter_columns_to_drop - list of columns to drop after filtering
+    :param fdedup_doc_column - document column name
+    :param fdedup_id_column - integer document id column name
+    :param fdedup_cluster_column - cluster column name
+    :param fdedup_bucket_cpu - number of CPUs per bucket hash
+    :param fdedup_doc_cpu - number of CPUs per doc hash
+    :param fdedup_mhash_cpu - number of CPUs per minhash hash
+    :param fdedup_num_permutations - number of permutations
+    :param fdedup_threshold - threshold
+    :param fdedup_shingles_size - number of words in shingle
+    :param fdedup_delimiters - delimiter for splitting document
+    :param fdedup_random_delay_limit - delay between reads to reduce S3 load.
+                                A random number between 0 and random_delay_limit is used
+    :param fdedup_snapshot_delay - delay between restoring individual actors
+    :param fdedup_use_bucket_snapshot - flag to skip buckets building and start from existing snapshots
+    :param fdedup_use_doc_snapshot - flag to skip documents building and start from existing snapshots
+    :param fdedup_n_samples - number of samples for parameters computation
     :return: None
     """
     # create clean_up task
@@ -118,8 +147,19 @@ def filtering(
         compute_exec_params = compute_exec_params_op(
             worker_options=ray_worker_options,
             actor_options=runtime_actor_options,
+            params={
+                "threshold": fdedup_threshold,
+                "num_permutations": fdedup_num_permutations,
+                "s3_config": data_s3_config,
+                "bucket_cpu": fdedup_bucket_cpu,
+                "doc_cpu": fdedup_doc_cpu,
+                "minhash_cpu": fdedup_mhash_cpu,
+            },
+            n_samples=fdedup_n_samples,
         )
         ComponentUtils.add_settings_to_component(compute_exec_params, ONE_HOUR_SEC * 2)
+        ComponentUtils.set_s3_env_vars_to_component(compute_exec_params, data_s3_access_secret)
+
         # start Ray cluster
         ray_cluster = create_ray_op(
             ray_name=ray_name,
@@ -131,7 +171,6 @@ def filtering(
         )
         ComponentUtils.add_settings_to_component(ray_cluster, ONE_HOUR_SEC * 2)
         ray_cluster.after(compute_exec_params)
-
         # Execute job
         execute_job = execute_ray_jobs_op(
             ray_name=ray_name,
@@ -141,21 +180,35 @@ def filtering(
                 "data_s3_config": data_s3_config,
                 "data_max_files": data_max_files,
                 "data_num_samples": data_num_samples,
-                "runtime_num_workers": compute_exec_params.output,
+                "runtime_num_workers": compute_exec_params.outputs["workers"],
                 "runtime_worker_options": runtime_actor_options,
                 "runtime_pipeline_id": runtime_pipeline_id,
                 "runtime_job_id": dsl.RUN_ID_PLACEHOLDER,
                 "runtime_code_location": runtime_code_location,
-                "filter_criteria_list": filter_criteria_list,
-                "filter_logical_operator": filter_logical_operator,
-                "filter_columns_to_drop": filter_columns_to_drop,
+                "fdedup_doc_column": fdedup_doc_column,
+                "fdedup_id_column": fdedup_id_column,
+                "fdedup_cluster_column": fdedup_cluster_column,
+                "fdedup_bucket_cpu": fdedup_bucket_cpu,
+                "fdedup_doc_cpu": fdedup_doc_cpu,
+                "fdedup_mhash_cpu": fdedup_mhash_cpu,
+                "fdedup_num_doc_actors": compute_exec_params.outputs["docs"],
+                "fdedup_num_bucket_actors": compute_exec_params.outputs["buckets"],
+                "fdedup_num_minhash_actors": compute_exec_params.outputs["min_hashes"],
+                "fdedup_num_preprocessors": compute_exec_params.outputs["preprocessors"],
+                "fdedup_num_permutations": fdedup_num_permutations,
+                "fdedup_threshold": fdedup_threshold,
+                "fdedup_shingles_size": fdedup_shingles_size,
+                "fdedup_delimiters": fdedup_delimiters,
+                "fdedup_random_delay_limit": fdedup_random_delay_limit,
+                "fdedup_snapshot_delay": fdedup_snapshot_delay,
+                "fdedup_use_doc_snapshot": fdedup_use_doc_snapshot,
+                "fdedup_use_bucket_snapshot": fdedup_use_bucket_snapshot,
             },
             exec_script_name=EXEC_SCRIPT_NAME,
             server_url=server_url,
         )
         ComponentUtils.add_settings_to_component(execute_job, ONE_WEEK_SEC)
         ComponentUtils.set_s3_env_vars_to_component(execute_job, data_s3_access_secret)
-
         execute_job.after(ray_cluster)
 
     # Configure the pipeline level to one week (in seconds)
@@ -164,4 +217,4 @@ def filtering(
 
 if __name__ == "__main__":
     # Compiling the pipeline
-    compiler.Compiler().compile(filtering, __file__.replace(".py", ".yaml"))
+    compiler.Compiler().compile(fdedup, __file__.replace(".py", ".yaml"))
