+++ transforms/universal/filter/ray/kfp-workflow/filter_wf.py
@@ -10,6 +10,8 @@
 # limitations under the License.
 ################################################################################
 
+# NOTE: This file is auto generated by Pipeline Generator.
+
 import kfp.compiler as compiler
 import kfp.components as comp
 import kfp.dsl as dsl
@@ -21,12 +23,16 @@ from kfp_support.workflow_support.utils import (
 
 
 # the name of the job script
-EXEC_SCRIPT_NAME: str = "malware_transform.py"
+EXEC_SCRIPT_NAME: str = "filter_transform.py"
+PREFIX: str = ""
 
-task_image = "quay.io/dataprep1/data-prep-kit/malware:0.4.0"
+task_image = "quay.io/dataprep1/data-prep-kit/filter:0.3.0"
 
 # components
 base_kfp_image = "quay.io/dataprep1/data-prep-kit/kfp-data-processing:0.1.0"
+
+# path to kfp component specifications files
+component_spec_path = "../../../../../kfp/kfp_ray_components/"
 
 # compute execution parameters. Here different tranforms might need different implementations. As
 # a result, insted of creating a component we are creating it in place here.
@@ -34,27 +40,30 @@ compute_exec_params_op = comp.func_to_container_op(
     func=ComponentUtils.default_compute_execution_params, base_image=base_kfp_image
 )
 # create Ray cluster
-create_ray_op = comp.load_component_from_file("../../../kfp_ray_components/createRayComponent.yaml")
+create_ray_op = comp.load_component_from_file(component_spec_path + "createRayComponent.yaml")
 # execute job
-execute_ray_jobs_op = comp.load_component_from_file("../../../kfp_ray_components/executeRayJobComponent.yaml")
+execute_ray_jobs_op = comp.load_component_from_file(component_spec_path + "executeRayJobComponent.yaml")
 # clean up Ray
-cleanup_ray_op = comp.load_component_from_file("../../../kfp_ray_components/cleanupRayComponent.yaml")
+cleanup_ray_op = comp.load_component_from_file(component_spec_path + "cleanupRayComponent.yaml")
+
 # Task name is part of the pipeline name, the ray cluster name and the job name in DMF.
-TASK_NAME: str = "malware"
+TASK_NAME: str = "filter"
 
 
+# Pipeline to invoke execution on remote resource
 @dsl.pipeline(
     name=TASK_NAME + "-ray-pipeline",
-    description="Pipeline for malware",
+    description="Pipeline for filtering task",
 )
-def malware(
-    ray_name: str = "malware-kfp-ray",  # name of Ray cluster
+def filtering(
+    # Ray cluster
+    ray_name: str = "filter-kfp-ray",  # name of Ray cluster
     ray_head_options: str = '{"cpu": 1, "memory": 4, "image_pull_secret": "", "image": "' + task_image + '" }',
     ray_worker_options: str = '{"replicas": 2, "max_replicas": 2, "min_replicas": 2, "cpu": 2, "memory": 4, '
     '"image_pull_secret": "", "image": "' + task_image + '"}',
     server_url: str = "http://kuberay-apiserver-service.kuberay.svc.cluster.local:8888",
     # data access
-    data_s3_config: str = "{'input_folder': 'test/malware/input', 'output_folder': 'test/malware/output'}",
+    data_s3_config: str = "{'input_folder': 'test/filter/input/', 'output_folder': 'test/filter/output/'}",
     data_s3_access_secret: str = "s3-secret",
     data_max_files: int = -1,
     data_num_samples: int = -1,
@@ -62,14 +71,15 @@ def malware(
     runtime_actor_options: str = "{'num_cpus': 0.8}",
     runtime_pipeline_id: str = "pipeline_id",
     runtime_code_location: str = "{'github': 'github', 'commit_hash': '12345', 'path': 'path'}",
-    # mallware
-    malware_input_column: str = "contents",
-    malware_output_column: str = "virus_detection",
+    # filtering parameters
+    filter_criteria_list: str = "['docq_total_words > 100 AND docq_total_words < 200', 'ibmkenlm_docq_perplex_score < 230']",
+    filter_logical_operator: str = "AND",
+    filter_columns_to_drop: str = "['extra', 'cluster']",
     # additional parameters
     additional_params: str = '{"wait_interval": 2, "wait_cluster_ready_tmout": 400, "wait_cluster_up_tmout": 300, "wait_job_ready_tmout": 400, "wait_print_tmout": 30, "http_retries": 5}',
 ):
     """
-    Pipeline to execute malware transform
+    Pipeline to execute Filtering transform
     :param ray_name: name of the Ray cluster
     :param ray_head_options: head node options, containing the following:
         cpu - number of cpus
@@ -91,16 +101,17 @@ def malware(
         wait_cluster_up_tmout - time to wait for cluster up, sec
         wait_job_ready_tmout - time to wait for job ready, sec
         wait_print_tmout - time between prints, sec
-        http_retries - httpt retries for API server calls
-    :param data_s3_config - s3 configuration
+        http_retries - http retries for API server calls
     :param data_s3_access_secret - s3 access secret
+    :param data_s3_config - s3 configuration
     :param data_max_files - max files to process
     :param data_num_samples - num samples to process
     :param runtime_actor_options - actor options
     :param runtime_pipeline_id - pipeline id
     :param runtime_code_location - code location
-    :param malware_input_column - input column name
-    :param malware_output_column - output column name
+    :param filter_criteria_list - list of filter criteria (in SQL WHERE clause format)
+    :param filter_logical_operator - logical operator (AND or OR) that joins filter criteria
+    :param filter_columns_to_drop - list of columns to drop after filtering
     :return: None
     """
     # create clean_up task
@@ -125,12 +136,12 @@ def malware(
         )
         ComponentUtils.add_settings_to_component(ray_cluster, ONE_HOUR_SEC * 2)
         ray_cluster.after(compute_exec_params)
+
         # Execute job
         execute_job = execute_ray_jobs_op(
             ray_name=ray_name,
             run_id=dsl.RUN_ID_PLACEHOLDER,
             additional_params=additional_params,
-            # note that the parameters below are specific for malware transform
             exec_params={
                 "data_s3_config": data_s3_config,
                 "data_max_files": data_max_files,
@@ -140,14 +151,16 @@ def malware(
                 "runtime_pipeline_id": runtime_pipeline_id,
                 "runtime_job_id": dsl.RUN_ID_PLACEHOLDER,
                 "runtime_code_location": runtime_code_location,
-                "malware_input_column": malware_input_column,
-                "malware_output_column": malware_output_column,
+                "filter_criteria_list": filter_criteria_list,
+                "filter_logical_operator": filter_logical_operator,
+                "filter_columns_to_drop": filter_columns_to_drop,
             },
             exec_script_name=EXEC_SCRIPT_NAME,
             server_url=server_url,
         )
         ComponentUtils.add_settings_to_component(execute_job, ONE_WEEK_SEC)
         ComponentUtils.set_s3_env_vars_to_component(execute_job, data_s3_access_secret)
+
         execute_job.after(ray_cluster)
 
     # Configure the pipeline level to one week (in seconds)
@@ -156,4 +169,4 @@ def malware(
 
 if __name__ == "__main__":
     # Compiling the pipeline
-    compiler.Compiler().compile(malware, __file__.replace(".py", ".yaml"))
+    compiler.Compiler().compile(filtering, __file__.replace(".py", ".yaml"))
